moved all build scripts to the build directory

+++ build-scripts/optimize.sh
@@ -2,6 +2,8 @@
 #
 # optimizes the project (css, javascript and images)
 
+buildscriptdir=`dirname "$0"`
+
 ####################################
 # remove require.js debugging code #
 ####################################
@@ -10,13 +12,17 @@ sed -i '/<script>/,/<\/script>/d' *.html || exit 1
 #####################################################
 # combine all images and write the image stylesheet #
 #####################################################
-./writesprite.sh || exit 1
+"$buildscriptdir"/writesprite.sh || exit 1
 
 #########################################
 # optimize and copy to ../boules-build/ #
 #########################################
+if ( "`readlink -f .`" == "`readlink -f ../boules-build`" ); then
+    echo "cannot remove ../boules-build when it's the current working directory">&2
+    exit 1
+fi
 rm -rf ../boules-build/ || exit 1
-r.js -o scripts/build.js || exit 1
+r.js -o "$buildscriptdir"/build.js || exit 1
 
 #########################################################################
 # delete the now-optimized files and copy them from the build directory #
